[refactor home page] update DI & router.
